+++ lib/dart_polisher.dart
@@ -1,4 +1,4 @@
-export 'src/dart_formatter.dart' hide BodyBitOpt;
+export 'src/dart_formatter.dart';
 export 'src/source_code.dart';
 export 'src/exceptions.dart';
 export 'src/style_fix.dart';